requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime-gpu>=1.16.0
tritonclient[all]>=2.40.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
